requirements.txt
requests
google-generativeai
python-dotenv
flask
flask_cors